Extract first round of common methods in Parser
