+++ src/main/java/seedu/duke/parsermanager/ParseDeleteClient.java
@@ -53,13 +53,6 @@ public class ParseDeleteClient extends Parser {
         }
     }
 
-    private void checkForEmptyDetails(String commandDetail) throws EmptyDetailException {
-        boolean isEmptyDetail = isEmptyString(commandDetail);
-        if (isEmptyDetail) {
-            throw new EmptyDetailException(EXCEPTION);
-        }
-    }
-
     private ArrayList<String> processCommandDetails(String rawCommandDetail)
             throws MissingFlagException, IncorrectFlagOrderException {
 
@@ -70,39 +63,11 @@ public class ParseDeleteClient extends Parser {
         return extractCommandDetails(rawCommandDetail, flags, flagIndexPositions);
     }
 
-    private boolean isEmptyString(String commandDetail) {
-        return commandDetail.trim().isEmpty();
-    }
-
-    private ArrayList<Integer> convertProcessedCommandDetailsToInteger(ArrayList<String> processedCommandDetail)
-            throws NotIntegerException {
-        ArrayList<Integer> integerDetails = new ArrayList<>();
-        for (String detail : processedCommandDetail) {
-            int integer;
-            try {
-                integer = Integer.parseInt(detail);
-            } catch (NumberFormatException e) {
-                throw new NotIntegerException(EXCEPTION);
-            }
-            integerDetails.add(integer - 1); // Convert to 0-index
-        }
-        return integerDetails;
-    }
-
     private void checkForInvalidClientIndexDelete(int clientIndex) throws InvalidIndexException {
         int currentListSize = clientList.getCurrentListSize();
         if (clientIndex < 0 || clientIndex >= currentListSize) {
             throw new InvalidIndexException(EXCEPTION);
         }
-    }
-
-    private int[] getFlagIndexPositions(String commandDetail, String[] flags) {
-        int[] flagIndexPositions = new int[flags.length];
-
-        for (int i = 0; i < flags.length; i++) {
-            flagIndexPositions[i] = commandDetail.indexOf(flags[i]);
-        }
-        return flagIndexPositions;
     }
 
     private void checkForMissingFlags(int[] flagIndexPositions) throws MissingFlagException {
@@ -119,27 +84,6 @@ public class ParseDeleteClient extends Parser {
         }
     }
 
-    private ArrayList<String> extractCommandDetails(String rawCommandDetail, String[] flags,
-                                                    int[] flagIndexPositions) {
-        ArrayList<String> extractedCommandDetails = new ArrayList<>();
-        for (int i = 0; i < flags.length; i++) {
-            String extractedDetail;
-            if (i == flags.length - 1) {
-                /* The extracted detail for the last flag starts from the char after the flag, to the end of
-                   rawCommandDetails */
-                extractedDetail = extractDetail(rawCommandDetail, flagIndexPositions[i] + flags[i].length());
-            } else {
-                // The extracted detail for non-last starts from the char after the flag, to index before the next flag
-                extractedDetail = extractDetail(
-                        rawCommandDetail,
-                        flagIndexPositions[i] + flags[i].length(),
-                        flagIndexPositions[i + 1]);
-            }
-            extractedCommandDetails.add(extractedDetail.trim());
-        }
-        return extractedCommandDetails;
-    }
-
     private boolean isFlagPresent(int flagIndexPosition) {
         return (flagIndexPosition != -1);
     }
@@ -150,14 +94,4 @@ public class ParseDeleteClient extends Parser {
             throw new IncorrectFlagOrderException(EXCEPTION);
         }
     }
-
-    private static String extractDetail(String rawDetail, int beginIndex) {
-        return rawDetail.substring(beginIndex).trim();
-    }
-
-    private static String extractDetail(String rawDetail, int beginIndex, int endIndex) {
-        return rawDetail.substring(beginIndex, endIndex).trim();
-    }
-
-
 }
