+++ src/main/java/seedu/duke/parsermanager/ParseDeleteProperty.java
@@ -51,13 +51,6 @@ public class ParseDeleteProperty extends Parser {
         }
     }
 
-    private void checkForEmptyDetails(String commandDetail) throws EmptyDetailException {
-        boolean isEmptyDetail = isStringEmpty(commandDetail);
-        if (isEmptyDetail) {
-            throw new EmptyDetailException(EXCEPTION);
-        }
-    }
-
     private ArrayList<String> processCommandDetails(String rawCommandDetail)
             throws MissingFlagException, IncorrectFlagOrderException {
 
@@ -68,39 +61,11 @@ public class ParseDeleteProperty extends Parser {
         return extractCommandDetails(rawCommandDetail, flags, flagIndexPositions);
     }
 
-    private ArrayList<Integer> convertProcessedCommandDetailsToInteger(ArrayList<String> processedCommandDetails)
-            throws NotIntegerException {
-        ArrayList<Integer> integerDetails = new ArrayList<>();
-        for (String detail : processedCommandDetails) {
-            int integer;
-            try {
-                integer = Integer.parseInt(detail);
-            } catch (NumberFormatException e) {
-                throw new NotIntegerException(EXCEPTION);
-            }
-            integerDetails.add(integer - 1); // Convert to 0-index
-        }
-        return integerDetails;
-    }
-
     private void checkForInvalidPropertyIndexDelete(int propertyIndex) throws InvalidIndexException {
         int currentListSize = propertyList.getCurrentListSize();
         if (propertyIndex < 0 || propertyIndex >= currentListSize) {
             throw new InvalidIndexException(EXCEPTION);
         }
-    }
-
-    private boolean isStringEmpty(String commandDetail) {
-        return commandDetail.trim().isEmpty();
-    }
-
-    private int[] getFlagIndexPositions(String commandDetails, String[] flags) {
-        int[] flagIndexPositions = new int[flags.length];
-
-        for (int i = 0; i < flags.length; i++) {
-            flagIndexPositions[i] = commandDetails.indexOf(flags[i]);
-        }
-        return flagIndexPositions;
     }
 
     private void checkForMissingFlags(int[] flagIndexPositions) throws MissingFlagException {
@@ -117,27 +82,6 @@ public class ParseDeleteProperty extends Parser {
         }
     }
 
-    private ArrayList<String> extractCommandDetails(String rawCommandDetail, String[] flags,
-                                                    int[] flagIndexPositions) {
-        ArrayList<String> extractedCommandDetails = new ArrayList<>();
-        for (int i = 0; i < flags.length; i++) {
-            String extractedDetail;
-            if (i == flags.length - 1) {
-                /* The extracted detail for the last flag starts from the char after the flag, to the end of
-                   rawCommandDetails */
-                extractedDetail = extractDetail(rawCommandDetail, flagIndexPositions[i] + flags[i].length());
-            } else {
-                // The extracted detail for non-last starts from the char after the flag, to index before the next flag
-                extractedDetail = extractDetail(
-                        rawCommandDetail,
-                        flagIndexPositions[i] + flags[i].length(),
-                        flagIndexPositions[i + 1]);
-            }
-            extractedCommandDetails.add(extractedDetail.trim());
-        }
-        return extractedCommandDetails;
-    }
-
     private boolean isFlagPresent(int flagIndexPosition) {
         return (flagIndexPosition != -1);
     }
@@ -147,13 +91,5 @@ public class ParseDeleteProperty extends Parser {
         if (!hasCorrectOrder) {
             throw new IncorrectFlagOrderException(EXCEPTION);
         }
-    }
-
-    private static String extractDetail(String rawDetail, int beginIndex) {
-        return rawDetail.substring(beginIndex).trim();
-    }
-
-    private static String extractDetail(String rawDetail, int beginIndex, int endIndex) {
-        return rawDetail.substring(beginIndex, endIndex).trim();
     }
 }
